net: wireless: bcmdhd: use static channel info pointer for each country
When country changed, it will made a new channel list in
same buffer. Sometimes different buffer pointer will have
same center frequence, which will made two Bss info
structures have different channel pointers but the channel's
center_freq is the same. So it is better to use static
channel pointers and use a flag to control channel valid
or not.
This is dhd modification adapt for the new wireless stack.

BUG=22398161

Change-Id: Ic84851980fe3842fc4856f75f08cb0b378a6dcdd

Signed-off-by: Xinrui <xinruis@broadcom.com>